feat: add NewProjectStore with form data management
