+++ src/renderer/components/_features/[start]/new-project/project-modal.tsx
@@ -1,55 +1,12 @@
 import { Modal, ModalContent, ModalTitle } from '@root/renderer/components/_molecules'
 import { useOpenPLCStore } from '@root/renderer/store'
-import { produce } from 'immer'
 import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { create } from 'zustand'
 
 import { Step1 } from './steps/first-step'
 import { Step2 } from './steps/second-step'
 import { Step3 } from './steps/third-step'
-
-export type NewProjectStoreProps = {
-  formData: {
-    type: string
-    name: string
-    path: string
-    language: string
-    time: string
-  }
-  setFormData: (form: NewProjectStoreProps['formData']) => void
-  resetFormData: () => void
-}
-
-const NewProjectStore = create<NewProjectStoreProps>((setState) => ({
-  formData: {
-    type: '',
-    name: '',
-    path: '',
-    language: '',
-    time: '',
-  },
-  setFormData: (form: Partial<NewProjectStoreProps['formData']>) => {
-    setState(
-      produce((state: NewProjectStoreProps) => {
-        Object.assign(state.formData, form)
-      }),
-    )
-  },
-  resetFormData: () => {
-    setState(
-      produce((state: NewProjectStoreProps) => {
-        state.formData = {
-          type: '',
-          name: '',
-          path: '',
-          language: '',
-          time: '',
-        }
-      }),
-    )
-  },
-}))
+import { NewProjectStore } from './store'
 
 const ProjectModal = ({ isOpen, onClose }: { isOpen: boolean; onClose: () => void }) => {
   const navigate = useNavigate()
@@ -59,6 +16,7 @@ const ProjectModal = ({ isOpen, onClose }: { isOpen: boolean; onClose: () => voi
   } = useOpenPLCStore()
   const resetFormData = NewProjectStore((state) => state.resetFormData)
   const [currentStep, setCurrentStep] = useState(1)
+
   const handleNext = () => {
     setCurrentStep((prevStep) => (prevStep < 3 ? prevStep + 1 : prevStep))
   }
@@ -82,6 +40,7 @@ const ProjectModal = ({ isOpen, onClose }: { isOpen: boolean; onClose: () => voi
     setEditingState('unsaved')
     clearTabs()
   }, [])
+
   const renderStep = () => {
     switch (currentStep) {
       case 1:
@@ -105,4 +64,4 @@ const ProjectModal = ({ isOpen, onClose }: { isOpen: boolean; onClose: () => voi
   )
 }
 
-export { NewProjectStore, ProjectModal }
+export { ProjectModal }
